pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "accent-ai-agent"
version = "0.1.0"
description = "Fast API backend for accent ai agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"fastapi[standard]<1.0.0,>=0.114.2",
	"pydantic>2.0",
	"jinja2<4.0.0,>=3.1.4",
	"sentry-sdk[fastapi]<2.0.0,>=1.40.6",
	"sqlmodel<1.0.0,>=0.0.21",
	"httpx<1.0.0,>=0.25.1",
	"psycopg[binary]<4.0.0,>=3.1.13",
	"pydantic-settings<3.0.0,>=2.2.1",
	"alembic<2.0.0,>=1.12.1",
	"tenacity<9.0.0,>=8.2.3",
	"ffmpeg-python>=0.2.0",
	"torchaudio>=2.7.0",
	"speechbrain>=1.0.3",
	"yt-dlp>=2025.5.22",
]

[tool.uv]
dev-dependencies = [
    "pytest<8.0.0,>=7.4.3",
    "mypy<2.0.0,>=1.8.0",
    "ruff<1.0.0,>=0.2.2",
    "pre-commit<4.0.0,>=3.6.2",
    "types-passlib<2.0.0.0,>=1.7.7.20240106",
    "coverage<8.0.0,>=7.4.3",
]

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic"]

[tool.ruff]
target-version = "py310"
exclude = ["alembic"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true
